requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
brotli==1.2.0
celery==5.5.3
certifi==2025.7.14
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
cssselect2==0.9.0
daphne==4.2.1
Django==4.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-unfold==0.89.0
djangorestframework==3.16.1
fonttools==4.62.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
Markdown==3.8.2
msgpack==1.1.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydyf==0.12.1
pyOpenSSL==25.3.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.4
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.1.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.11.0
zope.interface==8.0.1
zopfli==0.4.1
